workspace/codes/requirements.txt
psycopg2-binary==2.9.5
google-cloud-bigquery==3.6.0
pandas==1.5.3
pyarrow==11.0.0
db-dtypes==1.0.5